webkit/data/layout_test_results/v8/LayoutTests/fast/js/kde/string-2-n-expected.txt

CONSOLE MESSAGE: line 2: Uncaught SyntaxError: Unexpected token ILLEGAL
KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.successfullyParsed is undefined.

TEST COMPLETE
